feat: optimize performance with React.memo
- Add React.memo to SearchByRadarForm component
- Add React.memo to SelectedRadarsList component
- Create memoized RadarItem component for individual radar items
- Optimize re-rendering with strategic memoization
- Maintain all existing functionality and tests
