+++ src/__tests__/components/search-by-radar-form.test.tsx
@@ -4,12 +4,10 @@ import { useRouter } from 'next/navigation'
 import { SearchByRadarForm } from '@/app/(app)/veiculos/busca-por-radar/components/search-by-radar-form'
 import type { Radar } from '@/models/entities'
 
-// Mock do useRouter
 jest.mock('next/navigation', () => ({
   useRouter: jest.fn(),
 }))
 
-// Mock do contexto do mapa
 const mockSetSelectedObjects = jest.fn()
 const mockSetViewport = jest.fn()
 
@@ -19,7 +17,6 @@ jest.mock('@/hooks/useContexts/use-map-context', () => ({
   useMap: () => mockUseMap(),
 }))
 
-// Mock do hook de parâmetros de busca
 jest.mock('@/hooks/useParams/useCarRadarSearchParams', () => ({
   useCarRadarSearchParams: () => ({
     formattedSearchParams: {
@@ -64,7 +61,6 @@ describe('SearchByRadarForm', () => {
   })
 
   it('deve exibir lista de radares selecionados', () => {
-    // Mock com radares selecionados
     mockUseMap.mockReturnValue({
       layers: {
         radars: {
@@ -78,11 +74,9 @@ describe('SearchByRadarForm', () => {
 
     render(<SearchByRadarForm />)
 
-    // Abrir o popover
     const radarButton = screen.getByText('Radares (2)')
     fireEvent.click(radarButton)
 
-    // Verificar se os radares aparecem na lista
     expect(screen.getByText('0540461121')).toBeInTheDocument()
     expect(screen.getByText('0540461123')).toBeInTheDocument()
     expect(
@@ -91,7 +85,6 @@ describe('SearchByRadarForm', () => {
   })
 
   it('deve exibir contador correto de radares selecionados', () => {
-    // Mock com diferentes quantidades de radares
     mockUseMap.mockReturnValue({
       layers: {
         radars: {
@@ -105,12 +98,10 @@ describe('SearchByRadarForm', () => {
 
     render(<SearchByRadarForm />)
 
-    // Verificar se o contador está correto
     expect(screen.getByText('Radares (2)')).toBeInTheDocument()
   })
 
   it('deve renderizar o formulário sem erros', () => {
-    // Mock básico
     mockUseMap.mockReturnValue({
       layers: {
         radars: {
@@ -124,7 +115,6 @@ describe('SearchByRadarForm', () => {
 
     render(<SearchByRadarForm />)
 
-    // Verificar se o formulário renderiza
     expect(screen.getByText('Radares (0)')).toBeInTheDocument()
     expect(screen.getByText('Buscar')).toBeInTheDocument()
   })
